shifted the validate to MainLayout instead of App
