+++ src/App.tsx
@@ -1,45 +1,10 @@
 import MainLayout from "pages/MainLayout";
-import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import React from "react";
 import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
-import { AppDispatch, RootState } from "redux/store";
 import Login from "./pages/Login";
 import Register from "./pages/Register";
-import {
-  checkAuthorizedUser,
-  fetchLabels,
-  fetchNotes,
-  fetchNotesByLabels,
-} from "./redux/asyncThunks";
 
 function App() {
-  const dispatch = useDispatch<AppDispatch>();
-  const loggedInUserData = useSelector(
-    (state: RootState) => state.auth.loggedInUserData
-  );
-
-  useEffect(() => {
-    if (Object.keys(loggedInUserData).length !== 0) {
-      dispatch(fetchNotes());
-      dispatch(fetchLabels());
-      dispatch(fetchNotesByLabels());
-    } else {
-      dispatch(checkAuthorizedUser())
-        .unwrap()
-        .then(() => {
-          if (Object.keys(loggedInUserData).length !== 0) {
-          }
-        })
-        .catch((error) => {
-          console.error("Authorization failed:", error);
-        });
-    }
-  }, [dispatch, loggedInUserData]);
-
-  // useEffect(() => {
-  //   dispatch(refreshAccessToken());
-  // }, [dispatch]);
-
   return (
     <Router>
       <Routes>
